+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/room/RoomServiceExtensions.kt
@@ -14,6 +14,7 @@ import net.folivo.trixnity.core.model.events.StateEventContent
 import kotlin.jvm.JvmName
 
 @Deprecated("use getTimeline without roomId instead")
+@Suppress("DEPRECATION")
 fun RoomService.getTimeline(
     roomId: RoomId,
     onStateChange: suspend (TimelineStateChange<Flow<TimelineEvent>>) -> Unit = {},