+++ trixnity-clientserverapi/trixnity-clientserverapi-server/src/commonMain/kotlin/net/folivo/trixnity/clientserverapi/server/mediaApiRoutes.kt
@@ -10,16 +10,20 @@ internal fun Route.mediaApiRoutes(
     json: Json,
     contentMappings: EventContentSerializerMappings,
 ) {
+    @Suppress("DEPRECATION")
     matrixEndpoint(json, contentMappings, handler::getMediaConfigLegacy)
     matrixEndpoint(json, contentMappings, handler::getMediaConfig)
     matrixEndpoint(json, contentMappings, handler::createMedia)
     matrixEndpoint(json, contentMappings, handler::uploadMedia)
     matrixEndpoint(json, contentMappings, handler::uploadMediaByContentUri)
+    @Suppress("DEPRECATION")
     matrixEndpoint(json, contentMappings, handler::downloadMediaLegacy)
     matrixEndpoint(json, contentMappings, handler::downloadMedia)
     matrixEndpoint(json, contentMappings, handler::downloadMediaWithFileName)
+    @Suppress("DEPRECATION")
     matrixEndpoint(json, contentMappings, handler::downloadThumbnailLegacy)
     matrixEndpoint(json, contentMappings, handler::downloadThumbnail)
+    @Suppress("DEPRECATION")
     matrixEndpoint(json, contentMappings, handler::getUrlPreviewLegacy)
     matrixEndpoint(json, contentMappings, handler::getUrlPreview)
 }