+++ trixnity-client/src/commonTest/kotlin/net/folivo/trixnity/client/room/TimelineEventAggregationTest.kt
@@ -46,7 +46,7 @@ class TimelineEventAggregationTest : TrixnityBaseTest() {
         roomOutboxMessageStore = getInMemoryRoomOutboxMessageStore(),
         roomEventEncryptionServices = listOf(roomEventDecryptionServiceMock),
         mediaService = mediaServiceMock,
-        forgetRoomService = {},
+        forgetRoomService = { _, _ -> },
         userInfo = simpleUserInfo,
         timelineEventHandler = TimelineEventHandlerMock(),
         clock = testScope.testClock,